rebase to master and fix problems
