+++ frameworks/Go/goji/bash_profile.sh
@@ -1,0 +1,7 @@
+# Set the root of our go installation
+export GOROOT=${IROOT}/go
+
+# Where to find the go executable
+export PATH="$GOROOT/bin:$PATH"
+
+export GOPATH=${FWROOT}/go
