+++ frameworks/Go/goji/setup.py
@@ -1,0 +1,30 @@
+import subprocess
+import sys
+import os
+import setup_util
+
+def start(args, logfile, errfile):
+  setup_util.replace_text("goji/src/app/app.go", "tcp\(.*:3306\)", "tcp(" + args.database_host + ":3306)")
+  if os.name == 'nt':
+    #subprocess.call("rmdir /s /q pkg\\windows_amd64", shell=True, cwd="go")
+    #subprocess.call("rmdir /s /q src\\github.com", shell=True, cwd="go")
+    #subprocess.call("del /s /q /f bin\\hello.exe", shell=True, cwd="go")
+    subprocess.call("set GOPATH=C:\\FrameworkBenchmarks\\go&& go get ./...", shell=True, cwd="go", stderr=errfile, stdout=logfile)
+    subprocess.Popen("setup.bat", shell=True, cwd="go", stderr=errfile, stdout=logfile)
+    return 0
+  # os.environ["GOPATH"] = os.path.expanduser('~/FrameworkBenchmarks/go')
+  subprocess.call("go get ./...", shell=True, cwd="go", stderr=errfile, stdout=logfile)
+  subprocess.Popen("go run src/app/app.go src/app/models.go".rsplit(" "), cwd="go", stderr=errfile, stdout=logfile)
+  return 0
+def stop(logfile, errfile):
+  if os.name == 'nt':
+    subprocess.call("taskkill /f /im go.exe > NUL", shell=True, stderr=errfile, stdout=logfile)
+    subprocess.call("taskkill /f /im app.exe > NUL", shell=True, stderr=errfile, stdout=logfile)
+    return 0
+  p = subprocess.Popen(['ps', 'aux'], stdout=subprocess.PIPE)
+  out, err = p.communicate()
+  for line in out.splitlines():
+    if 'app' in line:
+      pid = int(line.split(None, 2)[1])
+      os.kill(pid, 15)
+  return 0
