+++ frameworks/Go/goji/templates/fortunes.tmpl
