+++ frameworks/Go/martini/templates/fortunes.tmpl
